+++ tests/test_flask_client.py
@@ -1,4 +1,3 @@
-
 import pytest
 
 import app as app_module